backend/pyproject.toml
[project]
name = "vidinsight-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.18.4",
    "anthropic>=0.84.0",
    "bcrypt>=4.2.0",
    "cryptography>=44.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "openai>=1.99.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn>=0.34.0",
    "youtube-transcript-api>=1.0.0",
]
